Break out the different policies.
