+++ pybo/bayesopt/policies/__init__.py
@@ -3,9 +3,15 @@ Acquisition functions.
 """
 
 # pylint: disable=wildcard-import
-from .simple import *
+from .improvement import *
+from .ucb import *
+from .thompson import *
 
-from . import simple
+from . import improvement
+from . import ucb
+from . import thompson
 
 __all__ = []
-__all__ += simple.__all__
+__all__ += improvement.__all__
+__all__ += ucb.__all__
+__all__ += thompson.__all__
